Generate regular grids

Creates a rectangular, hexagonal, or triangular grid of specified proportions. Returns tile vertex data formatted for geom_poygon()

Usage

make_tile_grid(
  grid = c("rect", "tri", "hex"),
  n = 40,
  direction = c("x", "y"),
  xlim,
  ylim
)

Arguments

grid

Character argument specifying geometry of grid to generate. Options include "rect" for rectangular grid (the default), "tri" for triangular grid, or "hex" for hexagonal grid.

n

Either a single integer specifying grid resolution in both dimensions, or a vector of length 2 specifying c(nx, ny) for different resolutions. Default is 40. Higher values create smoother surfaces but slower rendering.

direction

Either "x" (the default) or "y", specifying the orientation of tile rows. Ignored for rectangular grids.

xlim, ylim

Length-two numeric vectors bounding box over which to generate the grid.

Value

A data frame with the following columns: x, y, group (integer denoting) unique polygon id, and order (integer giving vertex order, for plotting).

Details

Grids are constructed such that tiles are approximately equilateral when scaled to a square domain, unless n gives separate resolution values for the two dimensions. For triangular and hexagonal grids, this means that n is only approximate.